Ensure currency is found during credit
Credit is called from Spree::Refund.process, which passes
gateway_options as {originator: self}. The only place currency is
obtainable from that is via the refund's (self).payment.gateway_options.
In case it is ever passed in correctly we are just going to
memoize it. Also adds the (empty) spree-required partial, which needs to be
named after the payment method.