Refactor redirection_controller confirm
To get HPP payments working, I have pulled some older code from other
forks of this project (watg/2-4-stable) and refactored the confirm
method in adyen_redirect_controller.rb. These changes will create the
payment at the confirm step of the process (at which time spree would
have already expected an unprocessed payment).

Additionally: This sets source_required to false for HPP
HPP payments will cause errors if source_required? is true. When set to
true, this causes the the authorize method to fire in adyen_hpp.rb,
which immediately raises an Adyen::HPPRedirectError. Setting flag to
false allows the payment to proceed to it's next state but will not save
the payment_source details. It is still possible to issue refunds with
this restriction however.